backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
pillow==10.1.0
python-dotenv==1.0.0
email-validator==2.1.0
argon2-cffi==23.1.0
